cron/requirements.txt
fastapi==0.101.1
uvicorn==0.23.2
fastapi-utils
websocket-client
requests